rename: myCommentInfoFilter.ts -> myInfoFilter.ts
